backend/clasificacionImagenes/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
google-generativeai
pillow
python-multipart
httpx
